Major: changed execution method
The previous way of doing executing code (feeding it into the
interpreter over stdin line by line) ended up being really finicky and
breaking easily. This replaces that by sending the whole code in as a
parameter and messing around to eliminate shared output later, though it
could be a bit worse if you're dealing with non-functional code (like
filesystem stuff, or anything with a random seed).